refactor:updated project folder
