+++ internal/services/factory.go
@@ -5,11 +5,11 @@ import (
 	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/config"
 	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/digitalOcean"
 	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/gcs"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/interfaces"
 	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/s3"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/pkg"
 )
 
-func NewUploader(cfg *config.Config) (interfaces.FileUploader, error) {
+func NewUploader(cfg *config.Config) (pkg.FileUploader, error) {
 	switch cfg.Provider {
 	case "gcs":
 		return gcs.NewGCSService(cfg.GCS.ProjectID, cfg.GCS.BucketName, cfg.GCS.CredentialsPath)
