+++ pkg/interfaces.go
@@ -1,4 +1,4 @@
-package interfaces
+package pkg
 
 import (
 	"context"
